light bulb pos fix - file organizations
