+++ Editor/Editor.vcxproj
@@ -110,10 +110,12 @@
     <ClCompile Include="Gizmo.cpp" />
     <ClCompile Include="GizmoPass.cpp" />
     <ClCompile Include="Grid.cpp" />
+    <ClCompile Include="LeftBar.cpp" />
     <ClCompile Include="MaterialView.cpp" />
     <ClCompile Include="MeshView.cpp" />
     <ClCompile Include="MultiChoiceParameterWindow.cpp" />
     <ClCompile Include="OutlinerWindow.cpp" />
+    <ClCompile Include="OverlayLighting.cpp" />
     <ClCompile Include="PluginWindow.cpp" />
     <ClCompile Include="PopupWindows.cpp" />
     <ClCompile Include="PrefabView.cpp" />
@@ -121,7 +123,10 @@
     <ClCompile Include="PublishManager.cpp" />
     <ClCompile Include="RenderSettingsView.cpp" />
     <ClCompile Include="SingleMaterialPass.cpp" />
+    <ClCompile Include="StatusBar.cpp" />
     <ClCompile Include="Thumbnail.cpp" />
+    <ClCompile Include="TopBar.cpp" />
+    <ClCompile Include="TopBar2d.cpp" />
     <ClCompile Include="TransformMod.cpp" />
     <ClCompile Include="EditorScene.cpp" />
     <ClCompile Include="UI.cpp" />
@@ -155,16 +160,21 @@
     <ClInclude Include="GizmoPass.h" />
     <ClInclude Include="Global.h" />
     <ClInclude Include="Grid.h" />
+    <ClInclude Include="LeftBar.h" />
     <ClInclude Include="MaterialView.h" />
     <ClInclude Include="MeshView.h" />
     <ClInclude Include="MultiChoiceParameterWindow.h" />
+    <ClInclude Include="OverlayLighting.h" />
     <ClInclude Include="PluginWindow.h" />
     <ClInclude Include="PopupWindows.h" />
     <ClInclude Include="PrefabView.h" />
     <ClInclude Include="PublishManager.h" />
     <ClInclude Include="RenderSettingsView.h" />
     <ClInclude Include="SingleMaterialPass.h" />
+    <ClInclude Include="StatusBar.h" />
     <ClInclude Include="Thumbnail.h" />
+    <ClInclude Include="TopBar.h" />
+    <ClInclude Include="TopBar2d.h" />
     <ClInclude Include="Workspace.h" />
     <ClInclude Include="OutlinerWindow.h" />
     <ClInclude Include="PropInspector.h" />
