+++ Editor/StatusBar.h
@@ -1,0 +1,17 @@
+#pragma once
+#include "OverlayUI.h"
+
+namespace ToolKit
+{
+  namespace Editor
+  {
+
+    class StatusBar : public OverlayUI
+    {
+     public:
+      explicit StatusBar(EditorViewport* owner);
+      void Show() override;
+    };
+
+  } // namespace Editor
+} // namespace ToolKit
